+++ pip_requ/scripts/_utils.py
@@ -1,0 +1,8 @@
+import subprocess
+
+DEVNULL = subprocess.DEVNULL
+
+
+def call(cmd, stdout=None, **kwargs):
+    formatted_cmd = [x.format(**kwargs) for x in cmd.split()]
+    return subprocess.check_call(formatted_cmd, stdout=stdout)