+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Unreleased
 
+- Add "pip-requ" command
+- Add "pip-requ build-wheels" command
+- Add "pip-requ compile-all" command
+- Add "pip-requ update" command
+
 # 0.180.0
 
 - Fork from pip-tools 1.8.0